+++ HISTORY.md
@@ -4,6 +4,11 @@ https://github.com/josdejong/workerpool
 
 ## not yet published, version 5.0.0
 
+- Deprecated option `nodeWorker` and created a new, more extensive option
+  `workerType` giving full control over the selected type of worker.
+  Added new option `'web'` to enforce use a Web Worker. See #85, #74.
+- In a node.js environment, the default `workerType` is changed from
+  `'process'` to `'thread'`.
 - Improved detection of environment (`browser` or `node`), fixing wrong
   detection in a Jest test environment. See #85.